feat: port entry 7 upstream loader changes (9b403861, 7d57f471)
Fix globalLoops field to read/store as int32 (not int16) in M2Loader
and SKELLoader. Remove UV V-coord transformation (v-1)*-1 from all
loaders (WMOLoader, WMOLegacyLoader, M2Loader, M2LegacyLoader,
M3Loader) to match JS source — UV stored as-is from file. Update
M2RendererGL gl_loops pointer type to int32_t to match.

+++ src/js/3D/loaders/M2Loader.h
@@ -138,7 +138,7 @@ public:
 	uint32_t flags = 0;
 	std::string name;
 
-	std::vector<int16_t> globalLoops;
+	std::vector<int32_t> globalLoops;
 	std::vector<M2Animation> animations;
 	std::vector<int16_t> animationLookup;
 	std::vector<M2Bone> bones;
